Move Sudo and OpSudoEvent to command package
