+++ src/main/java/org/betonquest/betonquest/BetonQuest.java
@@ -244,8 +244,8 @@ import org.betonquest.betonquest.quest.event.run.RunIndependentEventFactory;
 import org.betonquest.betonquest.quest.event.scoreboard.ScoreboardEventFactory;
 import org.betonquest.betonquest.quest.event.setblock.SetBlockEventFactory;
 import org.betonquest.betonquest.quest.event.stage.StageEventFactory;
-import org.betonquest.betonquest.quest.event.sudo.OpSudoEventFactory;
-import org.betonquest.betonquest.quest.event.sudo.SudoEventFactory;
+import org.betonquest.betonquest.quest.event.command.OpSudoEventFactory;
+import org.betonquest.betonquest.quest.event.command.SudoEventFactory;
 import org.betonquest.betonquest.quest.event.tag.TagGlobalEventFactory;
 import org.betonquest.betonquest.quest.event.tag.TagPlayerEventFactory;
 import org.betonquest.betonquest.quest.event.teleport.TeleportEventFactory;
