+++ src/main/java/org/betonquest/betonquest/quest/event/command/SudoEvent.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.quest.event.sudo;
+package org.betonquest.betonquest.quest.event.command;
 
 import org.betonquest.betonquest.VariableString;
 import org.betonquest.betonquest.api.profiles.Profile;
@@ -11,7 +11,7 @@ import java.util.List;
 /**
  * Forces the player to run commands.
  */
-public class OpSudoEvent implements Event {
+public class SudoEvent implements Event {
 
     /**
      * The commands to run.
@@ -23,19 +23,13 @@ public class OpSudoEvent implements Event {
      *
      * @param commands the commands to run
      */
-    public OpSudoEvent(final List<VariableString> commands) {
+    public SudoEvent(final List<VariableString> commands) {
         this.commands = commands;
     }
 
     @Override
     public void execute(final Profile profile) throws QuestRuntimeException {
         final Player player = profile.getOnlineProfile().get().getPlayer();
-        final boolean previousOp = player.isOp();
-        try {
-            player.setOp(true);
-            commands.forEach(command -> player.performCommand(command.getString(profile)));
-        } finally {
-            player.setOp(previousOp);
-        }
+        commands.forEach(command -> player.performCommand(command.getString(profile)));
     }
 }
