+++ src/main/java/org/betonquest/betonquest/quest/event/command/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.event.Event Event} implementation for command-related events.
+ */
+package org.betonquest.betonquest.quest.event.command;
