--- src/main/java/org/betonquest/betonquest/quest/event/sudo/package-info.java
@@ -1,4 +1,0 @@
-/**
- * {@link org.betonquest.betonquest.api.quest.event.Event Event} implementation for Sudo-related events.
- */
-package org.betonquest.betonquest.quest.event.sudo;
